# docs/openapi-v1.yaml
components:
  schemas:
...
    DraftCreateRequest:
      properties:
        agent:
          enum:
          - clu
          - tron
          - flynn
          title: Agent
          type: string
        body:
          minLength: 1
          title: Body
          type: string
        channel:
          default: imessage
          enum:
          - imessage
          - email
          title: Channel
          type: string
        in_reply_to_event_id:
          anyOf:
          - type: string
          - type: 'null'
          title: In Reply To Event Id
        preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview
        to_handle:
          minLength: 1
          title: To Handle
          type: string
      required:
      - agent
      - to_handle
      - body
      title: DraftCreateRequest
      type: object
    DraftListResponse:
      properties:
        drafts:
          items:
            $ref: '#/components/schemas/_DraftOut'
          title: Drafts
          type: array
      required:
      - drafts
      title: DraftListResponse
      type: object
    DraftPatchRequest:
      properties:
        approved_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Approved By
        body:
          anyOf:
          - type: string
          - type: 'null'
          title: Body
        reject_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reject Reason
        status:
          anyOf:
          - enum:
            - pending
            - approved
            - rejected
            - sent
            - send_failed
            type: string
          - type: 'null'
          title: Status
      title: DraftPatchRequest
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    IMessageSentRequest:
      properties:
        agent:
          enum:
          - clu
          - tron
          - flynn
          title: Agent
          type: string
        body:
          minLength: 1
          title: Body
          type: string
        error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Code
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        message_id:
          minLength: 1
          title: Message Id
          type: string
        sent_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Sent At
        status:
          enum:
          - success
          - failed
          title: Status
          type: string
        to:
          minLength: 1
          title: To
          type: string
      required:
      - agent
      - message_id
      - to
      - body
      - status
      title: IMessageSentRequest
      type: object
    IMessageSentResponse:
      properties:
        acknowledged:
          title: Acknowledged
          type: boolean
        event_id:
          title: Event Id
          type: string
      required:
      - acknowledged
      - event_id
      title: IMessageSentResponse
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
    _DraftOut:
      properties:
        agent:
          title: Agent
          type: string
        approved_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Approved At
        approved_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Approved By
        body:
          title: Body
          type: string
        channel:
          title: Channel
          type: string
        created_at:
          title: Created At
          type: string
        dispatch_message_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Dispatch Message Id
        draft_id:
          title: Draft Id
          type: string
        in_reply_to_event_id:
          anyOf:
          - type: string
          - type: 'null'
          title: In Reply To Event Id
        last_modified_at:
          title: Last Modified At
          type: string
        last_send_error_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Send Error Code
        last_send_error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Send Error Message
        preview:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview
        publisher:
          title: Publisher
          type: string
        reject_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reject Reason
        sent_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Sent At
        status:
          title: Status
          type: string
        to_handle:
          title: To Handle
          type: string
      required:
      - draft_id
      - agent
      - channel
      - to_handle
      - body
      - status
      - created_at
      - last_modified_at
      - publisher
      title: _DraftOut
      type: object
...
info:
  title: OpenClaw Bridge
  version: 1.0.0
openapi: 3.1.0
paths:
  /v1/agent/drafts:
    get:
      operationId: list_drafts_v1_agent_drafts_get
      parameters:
      - in: query
        name: agent
        required: false
        schema:
          anyOf:
          - enum:
            - clu
            - tron
            - flynn
            type: string
          - type: 'null'
          title: Agent
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - enum:
            - pending
            - approved
            - rejected
            - sent
            - send_failed
            type: string
          - type: 'null'
          title: Status
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DraftListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Drafts
      tags:
      - agent
    post:
      operationId: create_draft_v1_agent_drafts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftCreateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create Draft
      tags:
      - agent
  /v1/agent/drafts/{draft_id}:
    get:
      operationId: get_draft_v1_agent_drafts__draft_id__get
      parameters:
      - in: path
        name: draft_id
        required: true
        schema:
          minLength: 1
          title: Draft Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_DraftOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Draft
      tags:
      - agent
    patch:
      operationId: patch_draft_v1_agent_drafts__draft_id__patch
      parameters:
      - in: path
        name: draft_id
        required: true
        schema:
          minLength: 1
          title: Draft Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DraftPatchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/_DraftOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Patch Draft
      tags:
      - agent
...
  /v1/imessage/sent:
    post:
      operationId: imessage_sent_v1_imessage_sent_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IMessageSentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IMessageSentResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Imessage Sent
      tags:
      - imessage
...